Added support for reading in RDF from arbitrary sources and formats via the statements function.
Simply call statements on a String a File a URL or a Reader, grafter
will coerce the type.  If it's given a string grafter will try to coerce
it to a URL and if it is download and parse the statements from the
remote source, otherwise it will assume it's a file, and coerce it to a
file for reading.

Reading is done in a lazy manner, though to achieve this it has to hack
around deficiencies in the sesame parser API which means that it needs
more work to make it robust.