Create generalised lazy version of melt.
Create a generalised version of melt and implement melt and
melt-column-groups using it. The generalised melt takes a function
to partition the non-fixed columns and generates n rows in the output
for each row in the input, where n is the number of generated column
partitions. Rows in the output are generated by a function which
returns a map for the variable part of the row given the input row and
the generated column partition.

Melt is then implemented by returning n single-column partitions
and mapping each of these to a two-column fragment in the output
with 'value' and 'variable' columns.

Melt column groups is implemented by partitioning the columns into
n m-length partitions where m is the number of specified output
columns. The columns in each partition are then renamed to match
those specified.

Add tests for melt and melt-column-groups functions.