Remove variadic column name list from melt.
Remove the variadic column name list from the melt function and make
it an explicit single parameter. If a single column name is provided
it will be lifted into a collection, otherwise if multiple fixed
columns are required the caller must pass them in a (sequential)
collection.

This is a breaking API change.