Change read-dataset(s) to dispatch based on the data source type and format.
Change both read-dataset and read-datasets to dispatch first on the
data source type and secondly on the data format. The current approach
combines both data source type and data format in the dispatch function
and all implementations of the inner read-dataset(s)* functions dispatch
on the data format. Add new multi-methods read-dataset-source and
read-datasets-source which dispatch on the type of the data source. The
default implementations of both these functions attempt to infer the
data format from the source if it is not explicitly provided. This
inference process is defined by the DatasetFormat protocol and can be
extended. Once the format is known, the dispatch continues to the
existing read-dataset(s)* functions which dispatch based on the format.

At present all implementations of read-dataset(s)* assume the source
implements the IOFactory protocol and can be converted to a Reader or
InputStream, although this is not a requirement.