Allow optional key-value pairs to be provided to declare-pipeline.
Support an optional sequence of key-value pairs following the parameter
description map in declare-pipeline. These key-value pairs can be
used to provide additional options for the pipeline declaration.

Move the :supported-operation option from the parameter map to this
optional sequence.

Remove the additional arities for declare-pipeline and parse the
argument list manually. This is required since the optional
display-name parameter is in the middle of the parameter list.